pyproject.toml
[tool.poetry]
name = "njupt-smartclass-downloader"
version = "0.1.0"
description = ""
authors = ["vluna <vluna@qq.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.15" 
fastapi = "^0.128.0"
uvicorn = { extras = ["standard"], version = "^0.40.0" }
requests = "^2.32.5"
python-multipart = "^0.0.22"
jinja2 = "^3.1.6"
pycryptodome = "^3.23.0"
pytz = "^2025.2"
lxml = "^6.0.2"
nuitka = "^2.8.10"
#pyinstaller = "^6.18.0"
#flask = "^3.1.0"
#pyopenssl = "^24.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"